# Faster completion by splitting cache files