--- a/src/Entity/Cache.hs
+++ b/src/Entity/Cache.hs
@@ -12,34 +12,34 @@ import GHC.Generics
 data Cache = Cache
   { stmtList :: [Stmt.Stmt],
     remarkList :: [Remark],
-    locationTree :: LT.LocationTree,
-    localVarTree :: LVT.LocalVarTree,
-    topCandidate :: [TopCandidate],
-    rawImportSummary :: Maybe RawImportSummary
+    locationTree :: LT.LocationTree
   }
   deriving (Generic)
 
 data LowCache = LowCache
   { stmtList' :: [Stmt.StrippedStmt],
     remarkList' :: [Remark],
-    locationTree' :: LT.LocationTree,
-    localVarTree' :: LVT.LocalVarTree,
-    topCandidate' :: [TopCandidate],
-    rawImportSummary' :: Maybe RawImportSummary
+    locationTree' :: LT.LocationTree
   }
   deriving (Generic)
 
 instance Binary LowCache
+
+data CompletionCache = CompletionCache
+  { localVarTree :: LVT.LocalVarTree,
+    topCandidate :: [TopCandidate],
+    rawImportSummary :: Maybe RawImportSummary
+  }
+  deriving (Generic)
+
+instance Binary CompletionCache
 
 compress :: Cache -> LowCache
 compress cache =
   LowCache
     { stmtList' = map Stmt.compress (stmtList cache),
       remarkList' = remarkList cache,
-      locationTree' = locationTree cache,
-      localVarTree' = localVarTree cache,
-      topCandidate' = topCandidate cache,
-      rawImportSummary' = rawImportSummary cache
+      locationTree' = locationTree cache
     }
 
 extend :: LowCache -> Cache
@@ -47,8 +47,5 @@ extend cache =
   Cache
     { stmtList = map Stmt.extend (stmtList' cache),
       remarkList = remarkList' cache,
-      locationTree = locationTree' cache,
-      localVarTree = localVarTree' cache,
-      topCandidate = topCandidate' cache,
-      rawImportSummary = rawImportSummary' cache
+      locationTree = locationTree' cache
     }