This PR extracts completion-related parts from cache files. This results in about 5x faster completion.

## Example Case

Run `Scene.LSP.Complete.complete` at `(*)` in the following file:

```
define make-adder(x: int): (int) -> int {
  let adder =
    (y) => {
      add-int(x, y)
    }
  in
  adder
}

define main(): unit {
  print-int(make-adder(3)(7)); 
  // (*)
  print("\n")
}
```

## Result (Before)

```
	total time  =        0.06 secs   (88 ticks @ 1000 us, 10 processors)
	total alloc =  79,720,928 bytes  (excludes profiling overheads)

COST CENTRE             MODULE                     SRC                                         %time %alloc

get                     Entity.Hint                src/Entity/Hint.hs:(47,3)-(48,48)            68.2   43.7
get                     Entity.Term                src/Entity/Term.hs:44:10-39                  11.4    7.9
normalizeDir            Path.Posix                 src/Path/Include.hs:(782,1)-(790,25)          2.3    1.0
get                     Entity.PrimType            src/Entity/PrimType.hs:12:10-24               2.3    0.6
MAIN                    MAIN                       <built-in>                                    1.1    0.2
...
```

## Result (After)

```
	total time  =        0.01 secs   (19 ticks @ 1000 us, 10 processors)
	total alloc =  18,026,072 bytes  (excludes profiling overheads)

COST CENTRE                       MODULE                          SRC                                          %time %alloc

normalizeFilePath                 Path.Posix                      src/Path/Include.hs:(820,1)-(822,58)          26.3   21.0
lookup#                           Data.HashMap.Internal           Data/HashMap/Internal.hs:609:1-82             15.8    0.0
MAIN                              MAIN                            <built-in>                                    10.5    0.9
CAF                               GHC.IO.Handle.FD                <entire-module>                                5.3    0.3
takeWhile_                        Text.Megaparsec.Stream          Text/Megaparsec/Stream.hs:231:3-62             5.3    2.1
mplus                             Text.Megaparsec.Internal        Text/Megaparsec/Internal.hs:295:3-15           5.3    1.3
...
```